removing websockets-example-MNIST-parallel/run_websocket_server.py
